requirements.txt
cloudinary==1.8.0
CurrencyConverter==0.13.1
diff-match-patch==20121119
Django==1.11.6
django-ckeditor==5.4.0
django-cors-headers==2.1.0
django-filter==1.1.0
django-import-export==0.5.1
django-js-asset==0.1.1
django-sslserver==0.20
djangorestframework==3.7.1
et-xmlfile==1.0.1
factory-boy==2.9.2
Faker==0.8.6
image==1.5.16
jdcal==1.3
Markdown==2.6.9
mock==2.0.0
odfpy==1.3.5
olefile==0.44
openpyxl==2.4.9
pbr==3.1.1
Pillow==4.3.0
psycopg2==2.7.3.2
py-moneyed==0.7.0
PyPDF2==1.26.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
requests==2.18.4
six==1.11.0
tablib==0.12.1
text-unidecode==1.0
unicodecsv==0.14.1
urllib3==1.22
Wand==0.4.4
xlrd==1.1.0
xlwt==1.3.0
